Refactor munin::plugin to be generally more sane and simple.
This refactoring improves the DRYness of the define by setting source
and content on a base File instead of duplicating it.

It changes the parameter names according to the discussion around class
parameter naming, grouping the parameters by "target" and reusing the
naming from the underlying type.

It removes the gratuitous 'puppet:///modules/' prefix from the source
params, allowing local files and non-default puppet:// URLs as source.

I believe this to cover all previous use cases, and a few more.